feat: validation aturan hukuman jenis

+++ pencatatan-pelanggaran/app/Http/Controllers/Aturan/HukumanController.php
@@ -18,39 +18,96 @@ class HukumanController extends Controller
 
     public function store(Request $request)
     {
-        $id = Str::orderedUuid()->toString();
-
-        Hukuman::create([
-            'id' => $id,
-            'hukuman' => $request->hukuman,
-            $request->except(['_token'])
+        $request->validate([
+            'hukuman' => 'required'
         ]);
-
-        return redirect()->route('hukuman.index')->with('success','Data Berhasil Dibuat');
+        
+        try {
+            $id = Str::orderedUuid()->toString();
+    
+            Hukuman::create([
+                'id' => $id,
+                'hukuman' => $request->hukuman,
+                $request->except(['_token'])
+            ]);
+    
+            return redirect()
+                ->route('hukuman.index')
+                ->with('success','Data Berhasil Dibuat');
+            
+        } catch (\Throwable $th) {
+            return redirect()
+                ->route('hukuman.index')
+                ->with('error','Error Store Data');
+        }
     }
 
     public function edit(string $id)
     {
         $hukuman = Hukuman::find($id);
 
-        return view('home.admin.hukuman.edit', compact(['hukuman']));
+        if ($hukuman === null) {
+            return redirect()
+                ->route('hukuman.index')
+                ->with('success','Invalid Target Data');
+
+        } else {
+            return view('home.admin.hukuman.edit', compact(['hukuman']));
+        }
     }
 
     public function update(Request $request, string $id)
     {
         $hukuman = Hukuman::find($id);
 
-        $hukuman->update($request->except(['_token']));
+        if ($hukuman === null) {
+            return redirect()
+                ->route('hukuman.index')
+                ->with('error','Invalid Target Data');
+                
+        } else {
+            $request->validate([
+                'hukuman' => 'required'
+            ]);
 
-        return redirect()->route('hukuman.index')->with('success','Data Berhasil Diubah');
+            try {
+                $hukuman->update($request->except(['_token']));
+
+                return redirect()
+                    ->route('hukuman.index')
+                    ->with('success','Data Berhasil Diubah');
+    
+            } catch (\Throwable $th) {
+                return redirect()
+                    ->route('hukuman.index')
+                    ->with('error','Error Update Data');
+            }
+        }
+        
     }
 
     public function destroy(string $id)
     {
         $hukuman = Hukuman::find($id);
-
-        $hukuman->delete();
-
-        return redirect()->route('hukuman.index')->with('success','Data Berhasil Dihapus');
+        
+        if ($hukuman === null) {
+            return redirect()
+                ->route('hukuman.index')
+                ->with('error','Invalid Target Data');
+                
+        } else {
+            try {
+                $hukuman->delete();
+        
+                return redirect()
+                    ->route('hukuman.index')
+                    ->with('success','Data Berhasil Dihapus');
+            
+            } catch (\Throwable $th) {
+                return redirect()
+                    ->route('hukuman.index')
+                    ->with('error','Error Store Data');
+            }
+        }
     }
 }
